Commit before 4.22

+++ osa4/controllers/blog.js
@@ -3,14 +3,6 @@ const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs.map(blog => blog.toJSON()))
@@ -27,11 +19,14 @@ blogRouter.get('/', async (request, response) => {
   blogRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+    const poster = request.user
+    console.log(poster, "user in post")
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken || !decodedToken.id) {
       return response.status(401).json({error: 'token missing or invalid'})
     }
+    
     const user = await User.findById(decodedToken.id) 
   
     const blog = new Blog({
@@ -49,8 +44,22 @@ blogRouter.get('/', async (request, response) => {
   })
 
   blogRouter.delete('/:id', async (request, response, next) => {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString() ) {
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+    }
+
+    else return response.status(401).json({error: 'The blog does not match the user'})
   })
 
   blogRouter.put('/:id', async (request, response, next) => {
